Consolidate Karma configuration under karma/
