+++ karma/webrtc.conf.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const makeConf = require('./makeconf');
+const { join } = require('path');
+
+module.exports = makeConf(join('..', 'test', 'webrtc.js'));
